requirements.txt
pandas
openai
dotenv
streamlit
mysql-connector-python